+++ programs/locker/src/instructions/create_vesting_escrow.rs
@@ -52,7 +52,7 @@ impl CreateVestingEscrowParameters {
         &self,
         vesting_escrow: &AccountLoader<VestingEscrow>,
         recipient: Pubkey,
-        mint: Pubkey,
+        token_mint: Pubkey,
         sender: Pubkey,
         base: Pubkey,
         escrow_bump: u8,
@@ -69,7 +69,7 @@ impl CreateVestingEscrowParameters {
             self.amount_per_period,
             self.number_of_period,
             recipient,
-            mint,
+            token_mint,
             sender,
             base,
             escrow_bump,
