+++ programs/locker/src/instructions/mod.rs
@@ -1,4 +1,3 @@
-pub use cancel_vesting_escrow::*;
 pub use claim::*;
 pub use create_vesting_escrow::*;
 pub use create_vesting_escrow_metadata::*;
@@ -11,8 +10,6 @@ pub mod claim;
 pub mod create_vesting_escrow_metadata;
 
 pub mod update_vesting_escrow_recipient;
-
-pub mod cancel_vesting_escrow;
 
 pub mod v2;
