+++ tests/locker_utils/index.ts
@@ -273,80 +273,3 @@ export async function updateRecipient(params: UpdateRecipientParams) {
     }
   }
 }
-
-export interface CancelVestingPlanParams {
-  isAssertion: boolean;
-  tokenMint: web3.PublicKey;
-  escrow: web3.PublicKey;
-  rentReceiver: web3.PublicKey;
-  creatorToken: web3.PublicKey;
-  recipientToken: web3.PublicKey;
-  signer: web3.Keypair;
-}
-
-export async function cancelVestingPlan(
-  params: CancelVestingPlanParams,
-  claimable_amount: number,
-  total_amount: number
-) {
-  let {
-    isAssertion,
-    rentReceiver,
-    tokenMint,
-    escrow,
-    creatorToken,
-    recipientToken,
-    signer,
-  } = params;
-  const program = createLockerProgram(new Wallet(signer));
-  const escrowState = await program.account.vestingEscrow.fetch(escrow);
-
-  const escrowToken = getAssociatedTokenAddressSync(
-    escrowState.tokenMint,
-    escrow,
-    true,
-    TOKEN_PROGRAM_ID,
-    ASSOCIATED_TOKEN_PROGRAM_ID
-  );
-
-  const creator_token_balance_before = (
-    await program.provider.connection.getTokenAccountBalance(creatorToken)
-  ).value.amount;
-
-  const recipient_token_balance_before = (
-    await program.provider.connection.getTokenAccountBalance(recipientToken)
-  ).value.amount;
-
-  await program.methods
-    .cancelVestingEscrow()
-    .accounts({
-      escrow,
-      escrowToken,
-      creatorToken: creatorToken,
-      recipientToken: recipientToken,
-      tokenMint,
-      rentReceiver,
-      signer: signer.publicKey,
-      tokenProgram: TOKEN_PROGRAM_ID,
-      systemProgram: web3.SystemProgram.programId,
-    })
-    .rpc();
-
-  if (isAssertion) {
-    const escrowState = await program.account.vestingEscrow.fetch(escrow);
-
-    const creator_token_balance = (
-      await program.provider.connection.getTokenAccountBalance(creatorToken)
-    ).value.amount;
-    expect(
-      parseInt(creator_token_balance_before) + total_amount - claimable_amount
-    ).eq(parseInt(creator_token_balance));
-
-    const recipient_token_balance = (
-      await program.provider.connection.getTokenAccountBalance(recipientToken)
-    ).value.amount;
-    expect(parseInt(recipient_token_balance_before) + claimable_amount).eq(
-      parseInt(recipient_token_balance)
-    );
-  }
-}
